// src/app/models/event-interface.ts
export interface EventInterface{
    name?: string;
    place?: string;
    begin_date?: string;
    image?: File;
    start_hour?: string;
    final_date?: string;
    end_hour?: string;
    description?: string;
    organizer?: string;
    facebook?: string;
    instagram?: string;
    twitter?: string;
}